+++ Tests/CSharp/ExpressionTests/BinaryExpressionTests.cs
@@ -668,7 +668,7 @@ internal partial class TestClass712
     {
         bool? var1 = default;
         bool? var2 = default;
-        return (object)(var1.GetValueOrDefault() ? true : !var2 is not { } arg1 ? null : arg1 ? true : var1);
+        return var1.GetValueOrDefault() ? true : !var2 is not { } arg1 ? null : arg1 ? true : var1;
     }
 }");
     }