+++ Tests/TestData/MultiFileCharacterization/VBToCSResults/ConvertWholeSolution/VbNetStandardLib/My Project/MyNamespace.Static.2.Designer.cs
@@ -1624,7 +1624,7 @@ namespace Microsoft.VisualBasic
                 }
                 else if (Equals(reflectedType, typeof(string)))
                 {
-                    return (T)ToString(Value);
+                    return (T)(object)ToString(Value);
                 }
                 else
                 {